Recaive/requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.8
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
msgpack==1.1.0
numpy==2.0.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
pyqtgraph==0.13.7
python-can==4.5.0
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.12.2
wrapt==1.17.2
zipp==3.21.0
